requirements.txt
streamlit==1.34.0
ultralytics==8.3.222
segmentation-models-pytorch==0.5.0
albumentations==1.4.20
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
matplotlib==3.7.5
scikit-image==0.25.0
torch==2.5.1
torchvision==0.20.1
tensorflow==2.17.1
google-generativeai==0.8.3
keras==3.5.0
faiss-cpu==1.8.0
sentence-transformers==2.2.2
huggingface-hub==0.24.0
transformers==4.41.0
